WKWebView does not respect `-[UIScrollView allowsKeyboardScrolling]` https://bugs.webkit.org/show_bug.cgi?id=263164 rdar://115803459
Reviewed by Wenson Hsieh.

When determining if a keyboard scroll may occur, consult the value of `-[UIScrollView allowsKeyboardScrolling]`,
a new API introduced in iOS 17.

The default value of this property is `NO` on iPhone, and `YES` on iPad, Catalyst, and Vision. As such, the
property must initially be set to `YES` in WKScrollView to maintain its existing behavior.

Also adds an API test for this behavior.

* Source/WTF/wtf/PlatformHave.h:
* Source/WebKit/UIProcess/ios/WKContentViewInteraction.mm:
(-[WKContentView isScrollableForKeyboardScrollViewAnimator:]):
* Source/WebKit/UIProcess/ios/WKScrollView.mm:
(-[WKScrollView initWithFrame:]):
* Tools/TestWebKitAPI/Tests/ios/WKScrollViewTests.mm:
(TEST):

Canonical link: https://commits.webkit.org/269378@main