(REGRESSION (265731@main): fast/scrolling/rtl-scrollbars-animation-property.html is a constant failure) https://bugs.webkit.org/show_bug.cgi?id=262930 rdar://114349934
Reviewed by Cameron McCormack.

After https://commits.webkit.org/265731@main ScrollbarThemeMac::isLayoutDirectionRTL did not have an
NSScrollerImp to query about being in rtl mode. AppKit only uses userInterfaceLayoutDirection for
rectForPart/expandedRectForPart, which are not used with UI-side compositing, so for now if we detect
that the scrollbar should not be registered (due to UI-side compositing being on), just query the scrollable
area to see if the scrollbar is in rtl mode.

* Source/WebCore/platform/mac/ScrollbarThemeMac.mm:
(WebCore::ScrollbarThemeMac::isLayoutDirectionRTL):

Canonical link: https://commits.webkit.org/269390@main