An SVG element with a CSS reference filter fails to repaint when the filter changes https://bugs.webkit.org/show_bug.cgi?id=263229 rdar://117047658
Reviewed by Nikolas Zimmermann.

If an SVG element has a CSS filter which references an SVG element (which may be a child of that SVG,
or some other SVG in the document), then we fail to repaint the SVG when the filter changes.

This is a followup to 267236@main, removing the `&& !renderer().isSVGRootOrLegacySVGRoot()` check in
`RenderLayer::updateFiltersAfterStyleChange()`, since in this configuration we do need to have
the RenderLayer track the filter dependencies.

* LayoutTests/svg/filters/css-repaint-reference-filter-on-root-expected.txt: Added.
* LayoutTests/svg/filters/css-repaint-reference-filter-on-root.html: Added.
* Source/WebCore/rendering/RenderLayer.cpp:
(WebCore::RenderLayer::updateFiltersAfterStyleChange):

Canonical link: https://commits.webkit.org/269413@main