REGRESSION(266513@main?): [ macOS Monterey+ Debug WK1 ] media/modern-media-controls/tracks-support/text-track-selected-via-media-api.html is a constant timeout (259546) https://bugs.webkit.org/show_bug.cgi?id=263257 rdar://112952952
Reviewed by Jer Noble.

Changing WebKit1 test expectation for media/modern-media-controls/tracks-support/text-track-selected-via-media-api.html to "skip."
This test times out because it waits for the tracks button to appear in the bottom media controls bar, which never happens. The
test sets maximumRightContainerButtonCountOverride to 100, which should allow for the tracks button to be shown, but this does not
work for some reason, and so the maximumRightContainerButtonCount maintains its default value of 2, which only allows for AirPlay
and overflow buttons to appear. It does not feel important to fix this right now because maximumRightContainerButtonCountOverride
is only available in testing anyways, so the tracks button is never expected in any non-testing application.

* LayoutTests/platform/mac-wk1/TestExpectations:

Canonical link: https://commits.webkit.org/269437@main