Use-after-free in FullscreenManager::requestFullscreenForElement https://bugs.webkit.org/show_bug.cgi?id=263140 rdar://116736343
Reviewed by Chris Dumez.

Calling DeferredPromise::reject from the failedPreflights lambda in
FullscreenManager::requestFullscreenForElement may cause the Document that owns the
FullscreenManager to be deallocated, resulting in a use-after-free when the document is accessed
again after rejecting the promise. Resolved this by keeping a Ref to m_document for the lifetime of
the failedPreflights lambda.

Added a layout test.

* LayoutTests/fullscreen/fullscreen-cancel-after-request-crash-expected.txt: Added.
* LayoutTests/fullscreen/fullscreen-cancel-after-request-crash.html: Added.
* Source/WebCore/dom/FullscreenManager.cpp:
(WebCore::FullscreenManager::requestFullscreenForElement):

Canonical link: https://commits.webkit.org/267815.332@safari-7617-branch