Site isolated iframes should draw after navigation https://bugs.webkit.org/show_bug.cgi?id=263284 rdar://116496109
Reviewed by Tim Horton.

LocalFrameView setup needs to set the right size not only for main frames, but for root frames.
A root frame is the topmost LocalFrame in a process.  Before this change, a site-isolated iframe
would draw fine until you click on a link, and thereafter it would draw to a 0x0 LocalFrameView.
This updates the size properly, as verified by the layout test.

Many thanks to Matt Woodrow for investigating this after I got frustrated, being unable to find
what was going on myself.  He found what was going on, and this change is only a slight modification
of his original change that got it working.

Since scrolling site-isolated iframes doesn't work right now, this changed the expectations of
a layout test that scrolled a site-isolated iframe.  That is a future project.

* LayoutTests/http/tests/site-isolation/draw-after-navigation-expected.html: Added.
* LayoutTests/http/tests/site-isolation/draw-after-navigation.html: Added.
* LayoutTests/http/tests/site-isolation/resources/navigate-to-green-background.html: Added.
* Source/WebCore/page/LocalFrame.cpp:
(WebCore::LocalFrame::createView):

Canonical link: https://commits.webkit.org/269473@main