fast/events/touch/ios/selection-handles-after-touch-end.html is not robust against inflated (yet valid) selection handle views https://bugs.webkit.org/show_bug.cgi?id=263295 rdar://116418780
Reviewed by Wenson Hsieh.

The fast/events/touch/ios/selection-handles-after-touch-end.html test
makes exact assertions about the width of the text selection handle view
corresponding to the start grabber and the end grabber of a selection.
This means that the test is not robust against the case where the
grabberRect may visibly still have the correct width (2 px), but with a
view that has imperceptible inflated bounds.

We make the test robust to this case by asserting that the
startGrabberRect and endGrabberRect are roughly located around the left
edge and the right edge of the corresponding selection rect instead.

* LayoutTests/fast/events/touch/ios/selection-handles-after-touch-end-expected.txt:
* LayoutTests/fast/events/touch/ios/selection-handles-after-touch-end.html:

Canonical link: https://commits.webkit.org/269476@main